requirements.txt
annotated-types==0.6.0
anyio==4.3.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cors==1.0.1
dnspython==2.6.1
email_validator==2.1.1
exceptiongroup==1.2.1
fastapi==0.111.0
fastapi-cli==0.0.3
filelock==3.15.4
future==1.0.0
gevent==24.2.1
greenlet==3.0.3
h11==0.14.0
httpcore==1.0.5
httptools==0.6.1
httpx==0.27.0
idna==3.7
Jinja2==3.1.4
loguru==0.7.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
orjson==3.10.3
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.18.0
PySocks==1.7.1
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.32.3
requests-file==2.1.0
rich==13.7.1
shellingham==1.5.4
sniffio==1.3.1
starlette==0.37.2
tldextract==5.1.2
typer==0.12.3
typing_extensions==4.11.0
ujson==5.9.0
urllib3==2.2.2
uvicorn==0.29.0
uvloop==0.19.0
watchfiles==0.21.0
websockets==12.0
zope.event==5.0
zope.interface==6.4.post2
